Stage3/Devops/detector/requirements.txt
flask>=2.3.0        # Web framework for the live metrics dashboard
requests>=2.31.0    # HTTP client for sending Slack webhook notifications
pyyaml>=6.0         # Parses config.yaml into Python dicts
psutil>=5.9.0       # Reads CPU and memory usage for the dashboard
